Refactor ocr tasks.
