+++ sieves/tasks/__init__.py
@@ -10,14 +10,11 @@ from .predictive import (
     Translation,
 )
 from .predictive.core import PredictiveTask
-from .preprocessing import Chonkie, Docling, Marker, Unstructured
+from .preprocessing import Chonkie
 
 __all__ = [
     "Chonkie",
-    "Docling",
-    "Unstructured",
     "Classification",
-    "Marker",
     "InformationExtraction",
     "SentimentAnalysis",
     "Summarization",
