+++ sieves/tasks/preprocessing/ocr/__init__.py
@@ -1,3 +1,6 @@
-from sieves.tasks.preprocessing.ocr.core import OCR
+from .core import OCR
+from .docling_ import Docling
+from .marker_ import Marker
+from .unstructured_ import Unstructured
 
-__all__ = ["OCR"]
+__all__ = ["Docling", "Marker", "OCR", "Unstructured"]
